Move fitter to Actuators and rename to 'PolynomialPathFitter'
