+++ OpenSim/Actuators/RegisterTypes_osimActuators.cpp
@@ -77,7 +77,7 @@ OSIMACTUATORS_API void RegisterTypes_osimActuators()
     Object::RegisterType(MuscleFixedWidthPennationModel());
 
     Object::RegisterType(Millard2012EquilibriumMuscle());
-    Object::RegisterType(Millard2012AccelerationMuscle());        
+    Object::RegisterType(Millard2012AccelerationMuscle());
     Object::RegisterType(DeGrooteFregly2016Muscle());
 
     Object::registerType(ModelProcessor());
@@ -90,6 +90,8 @@ OSIMACTUATORS_API void RegisterTypes_osimActuators()
     Object::registerType(ModOpReplaceJointsWithWelds());
     Object::registerType(ModOpReplaceMusclesWithPathActuators());
     Object::registerType(ModOpReplacePathsWithFunctionBasedPaths());
+    Object::registerType(PolynomialPathFitter());
+    Object::registerType(PolynomialPathFitterBounds());
 
     //Object::RegisterType( ConstantMuscleActivation() );
     //Object::RegisterType( ZerothOrderMuscleActivationDynamics() );
@@ -107,7 +109,7 @@ OSIMACTUATORS_API void RegisterTypes_osimActuators()
     //Object::RenameType("Thelen2003Muscle", "Thelen2003Muscle_Deprecated");
 
   } catch (const std::exception& e) {
-    std::cerr 
+    std::cerr
         << "ERROR during osimActuators Object registration:\n"
         << e.what() << "\n";
   }
