+++ OpenSim/Actuators/osimActuators.h
@@ -45,6 +45,7 @@
 
 #include "ModelFactory.h"
 #include "ModelProcessor.h"
+#include "PolynomialPathFitter.h"
 
 #include "RegisterTypes_osimActuators.h"    // to expose RegisterTypes_osimActuators
